style/sign의 공용 태그 모듈화
