--- src/components/apis/userApis.js
@@ -1,5 +1,0 @@
-export const userApis = {
-  signUpUser: ({ id, password, passwordCheck }) => {},
-  signInUser: ({ id, password }) => {},
-  checkUserId: ({ username }) => {},
-};
