+++ src/lib/signStyle.js
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const StInput = styled.input`
+  width: 385px;
+  padding-left: 3px;
+  margin-bottom: 30px;
+  border: none;
+  background: none;
+  border-bottom: 2px solid black;
+  outline: none;
+`;
+
+export const StBtn = styled.button``;
+
+export const StBtnBox = styled.div``;
+
+export const StInputLabel = styled.p`
+  margin: 0;
+  font-weight: bold;
+  margin-bottom: 5px;
+`;
